adding new unit tests and updated functionality for agr-logger.
moved helper functions to agr-logger-helpers.  this functions are
intended to be private to agr-logger and not publicly exposed.  only
public api level functions should be defined in agr-logger.

also updated unit tests for agr-logger constructor.  more unit remain for
some of the new and updated constructor options.  the new agr-logger helper
function need unit tests as well.  some of those unit tests have been created
and added in this commit.